requirements.txt
Authlib==1.3.2
blinker==1.8.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
Flask==3.0.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonify==0.5
MarkupSafe==2.1.5
packaging==24.2
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.0.6
